# Conflicts:
#	src/main/java/willow/train/kuayue/event/client/TrainCouplerClientTickEvents.java
#	src/main/java/willow/train/kuayue/systems/train_extension/conductor/ConductorHelper.java

+++ src/main/java/willow/train/kuayue/network/s2c/TrainMergePacket.java
@@ -21,7 +21,7 @@ public class TrainMergePacket extends S2CPacket {
                 CreateClient.RAILWAYS.trains.get(buf.readUUID()),
                 buf.readBoolean(),
                 buf.readBoolean(),
-                buf.readInt(),
+                buf.readFloat(),
                 true
         );
     }
@@ -40,6 +40,6 @@ public class TrainMergePacket extends S2CPacket {
         buf.writeUUID(request.carriages().id);
         buf.writeBoolean(request.shouldReverseCarriages());
         buf.writeBoolean(request.isLocoHead());
-        buf.writeInt(request.spacing());
+        buf.writeFloat(request.spacing());
     }
 }
